Move ExcavatorHandler to separate package and extract MineralMix and MineralWorldInfo from it
